Created the new part files: _handler.dart, _deserialization.dart, _serialization.dart, and _utils.dart.
Created the logical handler classes: `Utils`, `TypeInfoHandler`, `SerializationHandler`, and `DeserializationHandler`.

Refactored the `JsonMapper` class to use the new handler classes and removed the obsolete `parts` directory.

Created the new focused utility files: `field_utils.dart`, `validation_utils.dart`, `reflection_utils.dart`, `json_utils.dart`, and `config_utils.dart`.

Created `reflection_handler.dart` and removed the old `reflection_utils.dart` file, centralizing all reflection logic.

Renamed the logic files to `field_inspector.dart`, `constraint_validator.dart`, `json_processor.dart`, and `options_handler.dart` and updated their class names.

Refactored and updated imports in `JsonMapper`, `SerializationHandler`, `DeserializationHandler`, and `TypeInfoHandler` to use the new focused logic files.

Created the barrel file `mapper/lib/mapper.dart` to provide a unified entry point to the library.

Updated all necessary files to use the new `FieldHandler` and `Json` classes, and removed the obsolete imports.

Created all the focused handler files: `reflection_handler.dart`, `field_handler.dart`, `serialization_handler.dart`, `deserialization_handler.dart`, and `type_info_handler.dart`.

Merged the logic from `JsonProcessor` and `OptionsHandler` into the `JsonMap` and `Json` classes, respectively, and removed the old files.

Refactored `JsonMapper` and updated all imports to use the new, consolidated logic handlers.

Created `field_handler.dart` with consolidated field inspection and validation logic, and added a newline.

Created `serialization_handler.dart` and `deserialization_handler.dart` with careful attention to imports and logic.

Refactored `JsonMapper` to use the new, consolidated logic handlers and updated all imports.

The `mapper.dart` file has been split into smaller, more manageable files, and the library is fully functional.
