+++ mapper/lib/src/mapper.dart
@@ -2,19 +2,18 @@ import 'dart:convert' show JsonEncoder, JsonDecoder;
 import 'dart:math';
 
 import 'package:collection/collection.dart' show IterableExtension;
-import 'package:reflectable/reflectable.dart'
-    show
-        ClassMirror,
-        InstanceMirror,
-        DeclarationMirror,
-        VariableMirror,
-        MethodMirror;
+import 'package:dart_json_mapper/src/model/index.dart';
+import 'package:reflectable/reflectable.dart' show InstanceMirror, MethodMirror;
 
-import 'errors.dart';
-import 'model/index.dart';
 import 'class_info.dart';
+import 'errors.dart';
 import 'globals.dart';
 import 'json_map.dart';
+import 'logic/deserialization_handler.dart';
+import 'logic/field_handler.dart';
+import 'logic/reflection_handler.dart';
+import 'logic/serialization_handler.dart';
+import 'logic/type_info_handler.dart';
 
 /// Singleton class providing mostly static methods for conversion of previously
 /// annotated by [JsonSerializable] Dart objects from / to JSON string
@@ -31,10 +30,10 @@ class JsonMapper {
   static String serialize(Object? object, [SerializationOptions? options]) {
     final context = SerializationContext(
         options ?? JsonMapper.globalSerializationOptions,
-        typeInfo: instance._getTypeInfo(object.runtimeType));
+        typeInfo: instance.typeInfoHandler.getTypeInfo(object.runtimeType));
     instance.clearCache();
-    return _getJsonEncoder(context)
-        .convert(instance._serializeObject(object, context));
+    return getJsonEncoder(context)
+        .convert(instance.serializationHandler.serializeObject(object, context));
   }
 
   /// Converts JSON [String] Or [Object] Or [Map<String, dynamic>] to Dart object instance of type T
@@ -52,16 +51,16 @@ class JsonMapper {
     assert(targetType != dynamic
         ? true
         : throw MissingTypeForDeserializationError());
-    return instance._deserializeObject(
+    return instance.deserializationHandler.deserializeObject(
         jsonValue != null
             ? jsonValue is String
-                ? _jsonDecoder.convert(jsonValue)
+                ? jsonDecoder.convert(jsonValue)
                 : jsonValue
             : null,
         DeserializationContext(targetOptions,
             classMeta:
-                instance._classes[targetType]?.getMeta(targetOptions.scheme),
-            typeInfo: instance._getTypeInfo(targetType))) as T?;
+                instance.classes[targetType]?.getMeta(targetOptions.scheme),
+            typeInfo: instance.typeInfoHandler.getTypeInfo(targetType))) as T?;
   }
 
   /// Converts Dart object to JSON String
@@ -70,7 +69,7 @@ class JsonMapper {
 
   /// Converts [getParams] object to Uri GET request with [baseUrl]
   static Uri toUri({Object? getParams, String? baseUrl = ''}) {
-    final params = _jsonDecoder
+    final params = jsonDecoder
         .convert(serialize(getParams))
         ?.entries
         .map((e) => '${e.key}=${Uri.encodeQueryComponent(e.value.toString())}')
@@ -87,9 +86,9 @@ class JsonMapper {
       [SerializationOptions? options]) {
     final context = SerializationContext(
         options ?? JsonMapper.globalSerializationOptions,
-        typeInfo: instance._getTypeInfo(object.runtimeType));
+        typeInfo: instance.typeInfoHandler.getTypeInfo(object.runtimeType));
     instance.clearCache();
-    final result = instance._serializeObject(object, context);
+    final result = instance.serializationHandler.serializeObject(object, context);
     return result is Map<String, dynamic> ? result : null;
   }
 
@@ -143,21 +142,21 @@ class JsonMapper {
   /// Adapters are meant to be used as a pluggable extensions, widening
   /// the number of supported types to be seamlessly converted to/from JSON
   JsonMapper useAdapter(IJsonMapperAdapter adapter, [int? priority]) {
-    if (_adapters.containsValue(adapter)) {
+    if (adapters.containsValue(adapter)) {
       return this;
     }
     final nextPriority = priority ??
-        (_adapters.keys.isNotEmpty
-            ? _adapters.keys.reduce((value, item) => max(value, item)) + 1
+        (adapters.keys.isNotEmpty
+            ? adapters.keys.reduce((value, item) => max(value, item)) + 1
             : 0);
-    _adapters[nextPriority] = adapter;
+    adapters[nextPriority] = adapter;
     _updateInternalMaps();
     return this;
   }
 
   /// De-registers previously registered adapter using [useAdapter] method
   JsonMapper removeAdapter(IJsonMapperAdapter adapter) {
-    _adapters.removeWhere((priority, x) => x == adapter);
+    adapters.removeWhere((priority, x) => x == adapter);
     _updateInternalMaps();
     return this;
   }
@@ -165,86 +164,95 @@ class JsonMapper {
   /// Prints out current mapper configuration to the console
   /// List of currently registered adapters and their priorities
   void info() =>
-      _adapters.forEach((priority, adapter) => print('$priority : $adapter'));
+      adapters.forEach((priority, adapter) => print('$priority : $adapter'));
 
   /// Wipes the internal caches
   void clearCache() {
-    _processedObjects.clear();
-    _convertedValuesCache.clear();
+    processedObjects.clear();
+    convertedValuesCache.clear();
   }
 
   static final JsonMapper instance = JsonMapper._internal();
-  static final JsonDecoder _jsonDecoder = JsonDecoder();
-  final _serializable = const JsonSerializable();
-  final Map<Type, ClassInfo> _classes = {};
-  final Map<String, ClassInfo> _mixins = {};
-  final Map<int, IJsonMapperAdapter> _adapters = {};
-  final Map<String, ProcessedObjectDescriptor> _processedObjects = {};
-  final Map<Type, ValueDecoratorFunction> _inlineValueDecorators = {};
-  final Map<Type, TypeInfo> _typeInfoCache = {};
-  final Map<dynamic, Type> _discriminatorToType = {};
+  static final JsonDecoder jsonDecoder = JsonDecoder();
+  final Map<Type, ClassInfo> classes = {};
+  final Map<String, ClassInfo> mixins = {};
+  final Map<int, IJsonMapperAdapter> adapters = {};
+  final Map<String, ProcessedObjectDescriptor> processedObjects = {};
+  final Map<Type, ValueDecoratorFunction> inlineValueDecorators = {};
+  final Map<Type, TypeInfo> typeInfoCache = {};
+  final Map<dynamic, Type> discriminatorToType = {};
   final Map<
           ICustomConverter?,
           Map<ConversionDirection,
               Map<DeserializationContext?, Map<dynamic, dynamic>>>>
-      _convertedValuesCache = {};
+      convertedValuesCache = {};
+
+  late final TypeInfoHandler typeInfoHandler;
+  late final SerializationHandler serializationHandler;
+  late final DeserializationHandler deserializationHandler;
 
   Map<Type, ICustomConverter> converters = {};
   Map<int, ITypeInfoDecorator> typeInfoDecorators = {};
   Map<Type, ValueDecoratorFunction> valueDecorators = {};
   Map<Type, dynamic> enumValues = {};
 
-  static JsonEncoder _getJsonEncoder(SerializationContext context) =>
+  static JsonEncoder getJsonEncoder(SerializationContext context) =>
       context.serializationOptions.indent != null &&
               context.serializationOptions.indent!.isNotEmpty
           ? JsonEncoder.withIndent(
-              context.serializationOptions.indent, _toEncodable(context))
-          : JsonEncoder(_toEncodable(context));
+              context.serializationOptions.indent, toEncodable(context))
+          : JsonEncoder(toEncodable(context));
 
-  static dynamic _toEncodable(SerializationContext context) =>
-      (Object? object) => instance._serializeObject(object, context);
+  static dynamic toEncodable(SerializationContext context) =>
+      (Object? object) => instance.serializationHandler.serializeObject(object, context);
 
   factory JsonMapper() => instance;
 
   JsonMapper._internal() {
+    typeInfoHandler = TypeInfoHandler(this);
+    serializationHandler = SerializationHandler(this);
+    deserializationHandler = DeserializationHandler(this);
     useAdapter(dartCoreAdapter);
     useAdapter(dartCollectionAdapter);
   }
 
   void _updateInternalMaps() {
-    _convertedValuesCache.clear();
-    _discriminatorToType.clear();
-    _typeInfoCache.clear();
-    _enumerateAnnotatedClasses((ClassInfo classInfo) {
-      final jsonMeta = classInfo.getMeta();
-      if (jsonMeta != null && jsonMeta.valueDecorators != null) {
-        _inlineValueDecorators.addAll(jsonMeta.valueDecorators!());
-      }
-      if (classInfo.reflectedType != null) {
-        _classes[classInfo.reflectedType!] = classInfo;
-      }
-      if (jsonMeta != null &&
-          jsonMeta.discriminatorValue != null &&
-          classInfo.reflectedType != null) {
-        _discriminatorToType.putIfAbsent(
-            jsonMeta.discriminatorValue, () => classInfo.reflectedType!);
-      }
-    });
+    convertedValuesCache.clear();
+    discriminatorToType.clear();
+    typeInfoCache.clear();
 
     enumValues = _enumValues;
     converters = _converters;
     typeInfoDecorators = _typeInfoDecorators;
     valueDecorators = _valueDecorators;
 
-    _enumerateAnnotatedClasses((ClassInfo classInfo) {
+    ReflectionHandler.enumerateAnnotatedClasses((classMirror) {
+      final classInfo = ClassInfo.fromCache(classMirror, classes);
+      final jsonMeta = classInfo.getMeta();
+      if (jsonMeta != null && jsonMeta.valueDecorators != null) {
+        inlineValueDecorators.addAll(jsonMeta.valueDecorators!());
+      }
+      if (classInfo.reflectedType != null) {
+        classes[classInfo.reflectedType!] = classInfo;
+      }
+      if (jsonMeta != null &&
+          jsonMeta.discriminatorValue != null &&
+          classInfo.reflectedType != null) {
+        discriminatorToType.putIfAbsent(
+            jsonMeta.discriminatorValue, () => classInfo.reflectedType!);
+      }
+    });
+
+    ReflectionHandler.enumerateAnnotatedClasses((classMirror) {
+      final classInfo = ClassInfo.fromCache(classMirror, classes);
       if (classInfo.superClass != null) {
         final superClassInfo =
-            ClassInfo.fromCache(classInfo.superClass!, _classes);
+            ClassInfo.fromCache(classInfo.superClass!, classes);
         final superClassTypeInfo = superClassInfo.reflectedType != null
-            ? _getTypeInfo(superClassInfo.reflectedType!)
+            ? typeInfoHandler.getTypeInfo(superClassInfo.reflectedType!)
             : null;
         if (superClassTypeInfo != null && superClassTypeInfo.isWithMixin) {
-          _mixins[superClassTypeInfo.mixinTypeName!] = classInfo;
+          mixins[superClassTypeInfo.mixinTypeName!] = classInfo;
         }
       }
     });
@@ -252,7 +260,7 @@ class JsonMapper {
 
   Map<Type, dynamic> get _enumValues {
     final result = <Type, dynamic>{};
-    for (var adapter in _adapters.values) {
+    for (var adapter in adapters.values) {
       result.addAll(adapter.enumValues);
     }
     return result;
@@ -260,7 +268,7 @@ class JsonMapper {
 
   Map<Type, ICustomConverter> get _converters {
     final result = <Type, ICustomConverter>{};
-    for (var adapter in _adapters.values) {
+    for (var adapter in adapters.values) {
       result.addAll(adapter.converters);
     }
     return result;
@@ -268,8 +276,8 @@ class JsonMapper {
 
   Map<Type, ValueDecoratorFunction> get _valueDecorators {
     final result = <Type, ValueDecoratorFunction>{};
-    result.addAll(_inlineValueDecorators);
-    for (var adapter in _adapters.values) {
+    result.addAll(inlineValueDecorators);
+    for (var adapter in adapters.values) {
       result.addAll(adapter.valueDecorators);
     }
     return result;
@@ -277,26 +285,16 @@ class JsonMapper {
 
   Map<int, ITypeInfoDecorator> get _typeInfoDecorators {
     final result = <int, ITypeInfoDecorator>{};
-    for (var adapter in _adapters.values) {
+    for (var adapter in adapters.values) {
       result.addAll(adapter.typeInfoDecorators);
     }
     return result;
   }
 
-  InstanceMirror? _safeGetInstanceMirror(Object object) {
-    InstanceMirror? result;
-    try {
-      result = _serializable.reflect(object);
-    } catch (error) {
-      return result;
-    }
-    return result;
-  }
-
-  String _getObjectKey(Object object) =>
+  String getObjectKey(Object object) =>
       '${object.runtimeType}-${identityHashCode(object)}';
 
-  ProcessedObjectDescriptor? _getObjectProcessed(Object object, int level) {
+  ProcessedObjectDescriptor? getObjectProcessed(Object object, int level) {
     ProcessedObjectDescriptor? result;
 
     if (object.runtimeType.toString() == 'Null' ||
@@ -304,125 +302,17 @@ class JsonMapper {
       return result;
     }
 
-    final key = _getObjectKey(object);
-    if (_processedObjects.containsKey(key)) {
-      result = _processedObjects[key];
+    final key = getObjectKey(object);
+    if (processedObjects.containsKey(key)) {
+      result = processedObjects[key];
       result!.logUsage(level);
     } else {
-      result = _processedObjects[key] = ProcessedObjectDescriptor(object);
+      result = processedObjects[key] = ProcessedObjectDescriptor(object);
     }
     return result;
   }
 
-  TypeInfo _getDeclarationTypeInfo(Type declarationType, Type? valueType) =>
-      _getTypeInfo((declarationType == dynamic && valueType != null)
-          ? valueType
-          : declarationType);
-
-  TypeInfo _getTypeInfo(Type type) {
-    if (_typeInfoCache[type] != null) {
-      return _typeInfoCache[type]!;
-    }
-    var result = TypeInfo(type);
-    for (var decorator in typeInfoDecorators.values) {
-      decorator.init(_classes, valueDecorators, enumValues);
-      result = decorator.decorate(result);
-    }
-    if (_mixins[result.typeName] != null) {
-      result.mixinType = _mixins[result.typeName]!.reflectedType;
-    }
-    _typeInfoCache[type] = result;
-    return result;
-  }
-
-  Type? _getGenericParameterTypeByIndex(
-          num parameterIndex, TypeInfo genericType) =>
-      genericType.isGeneric &&
-              genericType.parameters.length - 1 >= parameterIndex
-          ? genericType.parameters.elementAt(parameterIndex as int)
-          : null;
-
-  TypeInfo? _detectObjectType(dynamic objectInstance, Type? objectType,
-      JsonMap objectJsonMap, DeserializationContext context) {
-    final objectClassInfo = _classes[objectType];
-    if (objectClassInfo == null) {
-      return null;
-    }
-    final meta = objectClassInfo.getMeta(context.options.scheme);
-
-    if (objectInstance is Map<String, dynamic>) {
-      objectJsonMap = JsonMap(objectInstance, meta);
-    }
-    final typeInfo = _getTypeInfo(objectType ?? objectInstance.runtimeType);
-
-    final discriminatorProperty =
-        _getDiscriminatorProperty(objectClassInfo, context.options);
-    final discriminatorValue = discriminatorProperty != null &&
-            objectJsonMap.hasProperty(discriminatorProperty)
-        ? objectJsonMap.getPropertyValue(discriminatorProperty)
-        : null;
-    if (discriminatorProperty != null && discriminatorValue != null) {
-      final declarationMirror =
-          objectClassInfo.getDeclarationMirror(discriminatorProperty);
-      if (declarationMirror != null) {
-        final discriminatorType = _getDeclarationType(declarationMirror);
-        final value = _deserializeObject(discriminatorValue,
-            context.reBuild(typeInfo: _getTypeInfo(discriminatorType)));
-        if (value is Type) {
-          return _getTypeInfo(value);
-        }
-
-        if (_discriminatorToType[value] == null) {
-          final validDiscriminators = ClassInfo.getAllSubTypes(
-                  _classes, objectClassInfo)
-              .map((e) => e.getMeta(context.options.scheme)!.discriminatorValue)
-              .toList();
-          throw JsonMapperSubtypeError(
-            discriminatorValue,
-            validDiscriminators,
-            objectClassInfo,
-          );
-        }
-
-        return _getTypeInfo(_discriminatorToType[value]!);
-      }
-    }
-    if (discriminatorValue != null) {
-      final targetType = _getTypeByStringName(discriminatorValue);
-      return _classes[targetType] != null
-          ? _getTypeInfo(_classes[targetType]!.reflectedType!)
-          : typeInfo;
-    }
-    return typeInfo;
-  }
-
-  Type? _getTypeByStringName(String? typeName) =>
-      _classes.keys.firstWhereOrNull((t) => t.toString() == typeName);
-
-  Type _getDeclarationType(DeclarationMirror mirror) {
-    Type? result = dynamic;
-    VariableMirror variable;
-    MethodMirror method;
-
-    try {
-      variable = mirror as VariableMirror;
-      result = variable.hasReflectedType ? variable.reflectedType : null;
-    } catch (error) {
-      result = result;
-    }
-
-    try {
-      method = mirror as MethodMirror;
-      result =
-          method.hasReflectedReturnType ? method.reflectedReturnType : null;
-    } catch (error) {
-      result = result;
-    }
-
-    return result ??= dynamic;
-  }
-
-  ICustomConverter? _getConverter(
+  ICustomConverter? getConverter(
       JsonProperty? jsonProperty, TypeInfo typeInfo) {
     final result = jsonProperty?.converter ??
         converters[typeInfo.type!] ??
@@ -433,12 +323,12 @@ class JsonMapper {
 
     if (result is ICustomEnumConverter) {
       (result as ICustomEnumConverter)
-          .setEnumDescriptor(_getEnumDescriptor(enumValues[typeInfo.type!]));
+          .setEnumDescriptor(getEnumDescriptor(enumValues[typeInfo.type!]));
     }
     return result;
   }
 
-  IEnumDescriptor? _getEnumDescriptor(dynamic descriptor) {
+  IEnumDescriptor? getEnumDescriptor(dynamic descriptor) {
     if (descriptor is Iterable) {
       return EnumDescriptor(values: descriptor);
     }
@@ -448,40 +338,40 @@ class JsonMapper {
     return null;
   }
 
-  dynamic _getConvertedValue(ICustomConverter converter, dynamic value,
+  dynamic getConvertedValue(ICustomConverter converter, dynamic value,
       DeserializationContext context) {
     final direction = context.direction;
-    if (_convertedValuesCache.containsKey(converter) &&
-        _convertedValuesCache[converter]!.containsKey(direction) &&
-        _convertedValuesCache[converter]![direction]!.containsKey(context) &&
-        _convertedValuesCache[converter]![direction]![context]!
+    if (convertedValuesCache.containsKey(converter) &&
+        convertedValuesCache[converter]!.containsKey(direction) &&
+        convertedValuesCache[converter]![direction]!.containsKey(context) &&
+        convertedValuesCache[converter]![direction]![context]!
             .containsKey(value)) {
-      return _convertedValuesCache[converter]![direction]![context]![value];
+      return convertedValuesCache[converter]![direction]![context]![value];
     }
 
     final computedValue = direction == ConversionDirection.fromJson
         ? converter.fromJSON(value, context)
         : converter.toJSON(value, context as SerializationContext);
-    _convertedValuesCache.putIfAbsent(
+    convertedValuesCache.putIfAbsent(
         converter,
         () => {
               direction: {
                 context: {value: computedValue}
               }
             });
-    _convertedValuesCache[converter]!.putIfAbsent(
+    convertedValuesCache[converter]!.putIfAbsent(
         direction,
         () => {
               context: {value: computedValue}
             });
-    _convertedValuesCache[converter]![direction]!
+    convertedValuesCache[converter]![direction]!
         .putIfAbsent(context, () => {value: computedValue});
-    _convertedValuesCache[converter]![direction]![context]!
+    convertedValuesCache[converter]![direction]![context]!
         .putIfAbsent(value, () => computedValue);
     return computedValue;
   }
 
-  dynamic _applyValueDecorator(dynamic value, TypeInfo typeInfo) {
+  dynamic applyValueDecorator(dynamic value, TypeInfo typeInfo) {
     if (value == null) {
       return null;
     }
@@ -494,65 +384,9 @@ class JsonMapper {
     return value;
   }
 
-  bool _isNullableField(JsonProperty? meta) =>
-      !(JsonProperty.isRequired(meta) || JsonProperty.isNotNull(meta));
-
-  bool _isFieldIgnored(
-          [Json? classMeta,
-          JsonProperty? meta,
-          DeserializationOptions? options]) =>
-      (meta != null &&
-          (meta.ignore == true ||
-              ((meta.ignoreForSerialization == true ||
-                      JsonProperty.hasParentReference(meta) ||
-                      meta.inject == true) &&
-                  options is SerializationOptions) ||
-              (meta.ignoreForDeserialization == true &&
-                  options is! SerializationOptions)) &&
-          _isNullableField(meta));
-
-  bool _isFieldIgnoredByValue(
-          [dynamic value,
-          Json? classMeta,
-          JsonProperty? meta,
-          DeserializationOptions? options]) =>
-      ((meta != null &&
-              (_isFieldIgnored(classMeta, meta, options) ||
-                  meta.ignoreIfNull == true && value == null)) ||
-          (options is SerializationOptions &&
-              (((options.ignoreNullMembers == true ||
-                          classMeta?.ignoreNullMembers == true) &&
-                      value == null) ||
-                  ((_isFieldIgnoredByDefault(meta, classMeta, options)) &&
-                      JsonProperty.isDefaultValue(meta, value) == true)))) &&
-      _isNullableField(meta);
-
-  bool _isFieldIgnoredByDefault(
-          JsonProperty? meta, Json? classMeta, SerializationOptions options) =>
-      meta?.ignoreIfDefault == true ||
-      classMeta?.ignoreDefaultMembers == true ||
-      options.ignoreDefaultMembers == true;
-
-  void _enumerateAnnotatedClasses(Function visitor) {
-    for (var classMirror in _serializable.annotatedClasses) {
-      visitor(ClassInfo.fromCache(classMirror, _classes));
-    }
-  }
-
-  void _checkFieldConstraints(dynamic value, String name,
-      dynamic hasJsonProperty, JsonProperty? fieldMeta) {
-    if (JsonProperty.isNotNull(fieldMeta) &&
-        (hasJsonProperty == false || (value == null))) {
-      throw FieldCannotBeNullError(name, message: fieldMeta!.notNullMessage);
-    }
-    if (hasJsonProperty == false && JsonProperty.isRequired(fieldMeta)) {
-      throw FieldIsRequiredError(name, message: fieldMeta!.requiredMessage);
-    }
-  }
-
-  void _enumeratePublicProperties(InstanceMirror instanceMirror,
+  void enumeratePublicProperties(InstanceMirror instanceMirror,
       JsonMap? jsonMap, DeserializationContext context, Function visitor) {
-    final classInfo = ClassInfo.fromCache(instanceMirror.type, _classes);
+    final classInfo = ClassInfo.fromCache(instanceMirror.type, classes);
     final classMeta = classInfo.getMeta(context.options.scheme);
 
     for (var name in classInfo.publicFieldNames) {
@@ -560,21 +394,21 @@ class JsonMapper {
       if (declarationMirror == null) {
         continue;
       }
-      final declarationType = _getDeclarationType(declarationMirror);
+      final declarationType = ReflectionHandler.getDeclarationType(declarationMirror);
       final isGetterOnly = classInfo.isGetterOnly(name);
       final meta = classInfo.getDeclarationMeta(
           declarationMirror, context.options.scheme);
       if (meta == null &&
-          _getProcessAnnotatedMembersOnly(classMeta, context.options) == true) {
+          Json.getProcessAnnotatedMembersOnly(classMeta, context.options) == true) {
         continue;
       }
 
-      if (_isFieldIgnored(classMeta, meta, context.options)) {
+      if (FieldHandler.isFieldIgnored(classMeta, meta, context.options)) {
         continue;
       }
       final propertyContext =
           context.reBuild(classMeta: classMeta, jsonPropertyMeta: meta);
-      final property = _resolveProperty(
+      final property = resolveProperty(
           name, jsonMap, propertyContext, classMeta, meta, (name, jsonName, _) {
         var result = instanceMirror.invokeGetter(name);
         if (result == null && jsonMap != null) {
@@ -583,24 +417,24 @@ class JsonMapper {
         return result;
       });
 
-      _checkFieldConstraints(
+      FieldHandler.checkFieldConstraints(
           property.value, name, jsonMap?.hasProperty(property.name), meta);
 
-      if (_isFieldIgnoredByValue(
+      if (FieldHandler.isFieldIgnoredByValue(
           property.value, classMeta, meta, propertyContext.options)) {
         continue;
       }
       final typeInfo =
-          _getDeclarationTypeInfo(declarationType, property.value?.runtimeType);
-      visitor(name, property, isGetterOnly, meta, _getConverter(meta, typeInfo),
+          typeInfoHandler.getDeclarationTypeInfo(declarationType, property.value?.runtimeType);
+      visitor(name, property, isGetterOnly, meta, getConverter(meta, typeInfo),
           typeInfo);
     }
 
     classInfo.enumerateJsonGetters((MethodMirror mm, JsonProperty meta) {
-      final declarationType = _getDeclarationType(mm);
+      final declarationType = ReflectionHandler.getDeclarationType(mm);
       final propertyContext =
           context.reBuild(classMeta: classMeta, jsonPropertyMeta: meta);
-      final property = _resolveProperty(
+      final property = resolveProperty(
           mm.simpleName, jsonMap, propertyContext, classMeta, meta,
           (name, jsonName, _) {
         var result = instanceMirror.invoke(name, []);
@@ -610,20 +444,20 @@ class JsonMapper {
         return result;
       });
 
-      _checkFieldConstraints(property.value, mm.simpleName,
+      FieldHandler.checkFieldConstraints(property.value, mm.simpleName,
           jsonMap?.hasProperty(property.name), meta);
-      if (_isFieldIgnoredByValue(
+      if (FieldHandler.isFieldIgnoredByValue(
           property.value, classMeta, meta, context.options)) {
         return;
       }
       final typeInfo =
-          _getDeclarationTypeInfo(declarationType, property.value?.runtimeType);
+          typeInfoHandler.getDeclarationTypeInfo(declarationType, property.value?.runtimeType);
       visitor(mm.simpleName, property, true, meta,
-          _getConverter(meta, typeInfo), _getTypeInfo(declarationType));
+          getConverter(meta, typeInfo), typeInfoHandler.getTypeInfo(declarationType));
     }, context.options.scheme);
   }
 
-  PropertyDescriptor _resolveProperty(
+  PropertyDescriptor resolveProperty(
       String name,
       JsonMap? jsonMap,
       DeserializationContext context,
@@ -659,7 +493,6 @@ class JsonMapper {
         value = injectionJsonMap.getPropertyValue(jsonName);
         return PropertyDescriptor(jsonName, value, false);
       } else {
-        // TODO: Possibly would be better to throw an error when no injectable value provided?
         return PropertyDescriptor(jsonName, null, false);
       }
     }
@@ -669,165 +502,20 @@ class JsonMapper {
     }
     if (value == null &&
         meta?.defaultValue != null &&
-        _isFieldIgnoredByDefault(
+        FieldHandler.isFieldIgnoredByDefault(
             meta, classMeta, context.options as SerializationOptions)) {
       return PropertyDescriptor(jsonName!, meta?.defaultValue, true);
     }
     return PropertyDescriptor(jsonName!, value, true);
   }
 
-  void _enumerateConstructorParameters(ClassMirror classMirror, JsonMap jsonMap,
-      DeserializationContext context, Function filter, Function visitor) {
-    final classInfo = ClassInfo.fromCache(classMirror, _classes);
-    final classMeta = classInfo.getMeta(context.options.scheme);
-    final scheme =
-        classMeta != null ? classMeta.scheme : context.options.scheme;
-    final methodMirror = classInfo.getJsonConstructor(scheme);
-    if (methodMirror == null) {
-      return;
-    }
-    for (var param in methodMirror.parameters) {
-      if (!filter(param)) {
-        return;
-      }
-      final name = param.simpleName;
-      final declarationMirror = classInfo.getDeclarationMirror(name) ?? param;
-      final paramType = param.hasReflectedType
-          ? param.reflectedType
-          : param.hasDynamicReflectedType
-              ? param.dynamicReflectedType
-              : _getGenericParameterTypeByIndex(
-                      methodMirror.parameters.indexOf(param),
-                      context.typeInfo!) ??
-                  dynamic;
-      var paramTypeInfo = _getTypeInfo(paramType);
-      paramTypeInfo = paramTypeInfo.isDynamic
-          ? _getTypeInfo(_getDeclarationType(declarationMirror))
-          : paramTypeInfo;
-      final meta = classInfo.getDeclarationMeta(
-              declarationMirror, context.options.scheme) ??
-          classInfo.getDeclarationMeta(param, context.options.scheme);
-      final propertyContext = context.reBuild(
-          classMeta: classMeta,
-          jsonPropertyMeta: meta,
-          typeInfo: paramTypeInfo,
-          parentJsonMaps: <JsonMap>[
-            ...(context.parentJsonMaps ?? []),
-            jsonMap
-          ]);
-
-      // New logic to determine jsonNameForVisitor and finalValueForVisitor starts here:
-      dynamic finalValueForVisitor;
-      String? jsonNameForVisitor;
-
-      if (meta?.flatten == true) {
-        finalValueForVisitor = _deserializeObject(jsonMap.map, propertyContext.reBuild(jsonPropertyMeta: null));
-        jsonNameForVisitor = context.transformIdentifier(meta?.name ?? name);
-      } else { // This is the beginning of the 'else' block to be replaced
-        final property = _resolveProperty(
-            name,
-            jsonMap,
-            propertyContext,
-            classMeta,
-            meta,
-            (_, resolvedJsonNameFromCallback, defaultValueFromCallback) =>
-                jsonMap.hasProperty(resolvedJsonNameFromCallback)
-                    ? jsonMap.getPropertyValue(resolvedJsonNameFromCallback) ?? defaultValueFromCallback
-                    : defaultValueFromCallback);
-        jsonNameForVisitor = property.name;
-        if (property.raw) { // This check is crucial
-            finalValueForVisitor = _deserializeObject(property.value, propertyContext);
-        } else {
-            finalValueForVisitor = property.value;
-        }
-      } // This is the end of the 'else' block to be replaced
-
-      // Call the visitor with the determined jsonNameForVisitor and finalValueForVisitor
-      visitor(param, name, jsonNameForVisitor, classMeta, meta, finalValueForVisitor, paramTypeInfo);
-      // End of the new logic block for this section
-    }
-  }
-
-  String? _getDiscriminatorProperty(
-          ClassInfo classInfo, DeserializationOptions? options) =>
-      classInfo
-          .getMetaWhere((Json meta) => meta.discriminatorProperty != null,
-              options?.scheme)
-          ?.discriminatorProperty;
-
-  bool? _getProcessAnnotatedMembersOnly(
-          Json? meta, DeserializationOptions options) =>
-      meta != null && meta.processAnnotatedMembersOnly != null
-          ? meta.processAnnotatedMembersOnly
-          : options.processAnnotatedMembersOnly;
-
-  void _dumpDiscriminatorToObjectProperty(JsonMap object,
-      ClassMirror classMirror, DeserializationOptions? options) {
-    final classInfo = ClassInfo.fromCache(classMirror, _classes);
-    final discriminatorProperty = _getDiscriminatorProperty(classInfo, options);
-    if (discriminatorProperty != null) {
-      final typeInfo = _getTypeInfo(classMirror.reflectedType);
-      final lastMeta = classInfo.getMeta(options?.scheme);
-      final discriminatorValue =
-          (lastMeta != null && lastMeta.discriminatorValue != null
-                  ? lastMeta.discriminatorValue
-                  : typeInfo.typeName) ??
-              typeInfo.typeName;
-      object.setPropertyValue(discriminatorProperty, discriminatorValue);
-    }
-  }
-
-  Map<Symbol, dynamic> _getNamedArguments(
-      ClassMirror cm, JsonMap jsonMap, DeserializationContext context) {
-    final result = <Symbol, dynamic>{};
-
-    _enumerateConstructorParameters(
-        cm, jsonMap, context, (param) => param.isNamed, (param, name, jsonName,
-            classMeta, JsonProperty? meta, value, TypeInfo typeInfo) {
-      if (!_isFieldIgnoredByValue(value, classMeta, meta, context.options)) {
-        result[Symbol(name)] = value;
-      }
-    });
-
-    return result;
-  }
-
-  List _getPositionalArguments(ClassMirror cm, JsonMap jsonMap,
-      DeserializationContext context, List<String> positionalArgumentNames) {
-    final result = [];
-
-    _enumerateConstructorParameters(
-        cm, jsonMap, context, (param) => !param.isOptional && !param.isNamed,
-        (param, name, jsonName, classMeta, JsonProperty? meta, value,
-            TypeInfo typeInfo) {
-      positionalArgumentNames.add(name);
-      result.add(_isFieldIgnoredByValue(value, classMeta, meta, context.options)
-          ? null
-          : value);
-    });
-
-    return result;
-  }
-
-  bool _isValidJSON(dynamic jsonValue) {
-    try {
-      if (jsonValue is String) {
-        _jsonDecoder.convert(jsonValue);
-        return true;
-      }
-      return false;
-    } on FormatException {
-      return false;
-    }
-  }
-
-  void _configureConverter(
+  void configureConverter(
       ICustomConverter converter, DeserializationContext context,
       {dynamic value}) {
     if (converter is ICompositeConverter) {
-      (converter as ICompositeConverter).setGetConverterFunction(_getConverter);
+      (converter as ICompositeConverter).setGetConverterFunction(getConverter);
       (converter as ICompositeConverter)
-          .setGetConvertedValueFunction(_getConvertedValue);
+          .setGetConvertedValueFunction(getConvertedValue);
     }
     if (converter is ICustomIterableConverter) {
       (converter as ICustomIterableConverter).setIterableInstance(value);
@@ -838,300 +526,10 @@ class JsonMapper {
     }
     if (converter is IRecursiveConverter) {
       (converter as IRecursiveConverter).setSerializeObjectFunction(
-          (o, context) => _serializeObject(o, context));
+          (o, context) => serializationHandler.serializeObject(o, context));
       (converter as IRecursiveConverter).setDeserializeObjectFunction((o,
               context, type) =>
-          _deserializeObject(o, context.reBuild(typeInfo: _getTypeInfo(type))));
+          deserializationHandler.deserializeObject(o, context.reBuild(typeInfo: typeInfoHandler.getTypeInfo(type))));
     }
-  }
-
-  dynamic _serializeObject(Object? object, SerializationContext? context) {
-    if (object == null) {
-      return object;
-    }
-
-    final im = _safeGetInstanceMirror(object);
-    final converter = _getConverter(
-        context!.jsonPropertyMeta, _getTypeInfo(object.runtimeType));
-    if (converter != null) {
-      _configureConverter(converter, context, value: object);
-      return _getConvertedValue(converter, object, context);
-    }
-
-    if (im == null) {
-      if (context.serializationOptions.ignoreUnknownTypes == true) {
-        return null;
-      } else {
-        throw MissingAnnotationOnTypeError(object.runtimeType);
-      }
-    }
-
-    final classInfo = ClassInfo.fromCache(im.type, _classes);
-    final jsonMeta = classInfo.getMeta(context.options.scheme);
-    final initialMap = context.level == 0
-        ? context.options.template ?? <String, dynamic>{}
-        : <String, dynamic>{};
-    final result = JsonMap(initialMap, jsonMeta);
-    final processedObjectDescriptor =
-        _getObjectProcessed(object, context.level);
-    if (processedObjectDescriptor != null &&
-        processedObjectDescriptor.levelsCount > 1) {
-      final allowanceIsSet =
-          (jsonMeta != null && jsonMeta.allowCircularReferences! > 0);
-      final allowanceExceeded = (allowanceIsSet &&
-              processedObjectDescriptor.levelsCount >
-                  jsonMeta.allowCircularReferences!)
-          ? true
-          : null;
-      if (allowanceExceeded == true) {
-        return null;
-      }
-      if (allowanceIsSet == false) {
-        throw CircularReferenceError(object);
-      }
-    }
-    _dumpDiscriminatorToObjectProperty(result, im.type, context.options);
-    _enumeratePublicProperties(im, null, context, (name, property, isGetterOnly,
-        JsonProperty? meta, converter, TypeInfo typeInfo) {
-      dynamic convertedValue;
-      final propertyContext = context.reBuild(
-          level: context.level + 1,
-          jsonPropertyMeta: meta,
-          classMeta: jsonMeta, // jsonMeta is from _serializeObject scope
-          typeInfo: typeInfo) as SerializationContext;
-
-  if (meta?.rawJson == true && typeInfo.type == String && property.value is String) {
-    final jsonString = property.value as String;
-    if (jsonString.isEmpty || jsonString == "null") {
-      convertedValue = null;
-    } else {
-      try {
-        convertedValue = JsonDecoder().convert(jsonString);
-      } on FormatException {
-        convertedValue = jsonString; // Treat as plain string
-      }
-    }
-  } else { // Not a rawJson string field
-    if (meta?.flatten == true) {
-      final Map flattenedPropertiesMap =
-          _serializeObject(property.value, propertyContext);
-      final fieldPrefixWords = meta?.name != null
-          ? toWords(meta?.name, propertyContext.caseStyle).join(' ')
-          : null;
-      for (var element in flattenedPropertiesMap.entries) {
-        result.setPropertyValue(
-            fieldPrefixWords != null
-                ? transformIdentifierCaseStyle(
-                    transformIdentifierCaseStyle(
-                        '$fieldPrefixWords ${element.key}',
-                        defaultCaseStyle,
-                        null),
-                    propertyContext.targetCaseStyle,
-                    defaultCaseStyle)
-                : element.key,
-            element.value);
-      }
-      return; // from callback
-    }
-
-    final actualConverter = _getConverter(meta, typeInfo); // Use _getConverter here
-    if (actualConverter != null) {
-      final valueToConvert = property.value ?? meta?.defaultValue;
-      _configureConverter(actualConverter, propertyContext, value: valueToConvert);
-      convertedValue = _getConvertedValue(actualConverter, valueToConvert, propertyContext);
-    } else {
-      // No specific converter found for the type (e.g. custom object, non-string primitive not covered)
-      convertedValue = _serializeObject(property.value, propertyContext);
-    }
-  }
-  result.setPropertyValue(property.name, convertedValue ?? meta?.defaultValue);
-    });
-
-    final jsonAnyGetter = classInfo.getJsonAnyGetter();
-    if (jsonAnyGetter != null) {
-      final anyMap = im.invoke(jsonAnyGetter.simpleName, [])!;
-      result.map.addAll(anyMap as Map<String, dynamic>);
-    }
-
-    return result.map;
-  }
-
-  Object? _deserializeObject(
-      dynamic jsonValue, DeserializationContext context) {
-    if (jsonValue == null) {
-      return null;
-    }
-    var typeInfo = context.typeInfo!;
-    final converter = _getConverter(context.jsonPropertyMeta, typeInfo);
-    if (converter != null) {
-      _configureConverter(converter, context);
-      if (typeInfo.isIterable &&
-          (converter is ICustomIterableConverter &&
-              converter is! DefaultIterableConverter)) {
-        return _applyValueDecorator(
-            _getConvertedValue(converter, jsonValue, context), typeInfo);
-      }
-      return _applyValueDecorator(
-          _getConvertedValue(converter, jsonValue, context), typeInfo);
-    }
-
-    dynamic convertedJsonValue =
-        _isValidJSON(jsonValue) ? _jsonDecoder.convert(jsonValue) : jsonValue;
-
-    if (convertedJsonValue is String && typeInfo.type is Type) {
-      return _getTypeByStringName(convertedJsonValue.replaceAll("\"", ""));
-    }
-
-    final jsonMap = JsonMap(
-        convertedJsonValue, null, context.parentJsonMaps as List<JsonMap>?);
-    typeInfo =
-        _detectObjectType(null, context.typeInfo!.type, jsonMap, context) ??
-            typeInfo;
-    final classInfo = _classes[typeInfo.type] ??
-        _classes[typeInfo.genericType] ??
-        _classes[typeInfo.mixinType];
-    if (classInfo == null) {
-      throw MissingAnnotationOnTypeError(typeInfo.type);
-    }
-    jsonMap.jsonMeta = classInfo.getMeta(context.options.scheme);
-
-    final namedArguments =
-        _getNamedArguments(classInfo.classMirror, jsonMap, context);
-    final positionalArgumentNames = <String>[];
-    final positionalArguments = _getPositionalArguments(
-        classInfo.classMirror, jsonMap, context, positionalArgumentNames);
-    dynamic objectInstance;
-    try {
-      objectInstance = context.options.template ??
-          (classInfo.classMirror.isEnum
-              ? null
-              : classInfo.classMirror.newInstance(
-                  classInfo
-                      .getJsonConstructor(context.options.scheme)!
-                      .constructorName,
-                  positionalArguments,
-                  namedArguments));
-    } on TypeError catch (typeError) {
-      final positionalNullArguments = positionalArgumentNames.where((element) =>
-          positionalArguments[positionalArgumentNames.indexOf(element)] ==
-          null);
-      final namedNullArguments = Map<Symbol, dynamic>.from(namedArguments);
-      namedNullArguments.removeWhere((key, value) => value != null);
-      throw CannotCreateInstanceError(
-          typeError, classInfo, positionalNullArguments, namedNullArguments);
-    }
-
-    final im = _safeGetInstanceMirror(objectInstance)!;
-    final inheritedPublicFieldNames = classInfo.inheritedPublicFieldNames;
-    final mappedFields = namedArguments.keys
-        .map((Symbol symbol) =>
-            RegExp('"(.+)"').allMatches(symbol.toString()).first.group(1))
-        .toList()
-      ..addAll(positionalArgumentNames);
-
-    _enumeratePublicProperties(im, jsonMap, context, (name, property,
-        isGetterOnly, JsonProperty? meta, converter, TypeInfo typeInfo) {
-      final propertyContext = context.reBuild(
-          parentObjectInstances: [
-            ...(context.parentObjectInstances ?? []),
-            objectInstance
-          ],
-          typeInfo: typeInfo,
-          jsonPropertyMeta: meta,
-          parentJsonMaps: <JsonMap>[
-            ...(context.parentJsonMaps ?? []),
-            jsonMap
-          ]);
-      final defaultValue = meta?.defaultValue;
-      final hasJsonProperty = jsonMap.hasProperty(property.name);
-      var fieldValue = jsonMap.getPropertyValue(property.name);
-      if (!hasJsonProperty || mappedFields.contains(name)) {
-      // BEGINNING OF BLOCK TO REPLACE/OVERWRITE
-      if (meta?.flatten == true) {
-          // name: the simple name of the property (e.g., "page")
-          // mappedFields: list of names of fields handled by constructor
-          // isGetterOnly: boolean, true if the property is final or getter-only
-
-          if (mappedFields.contains(name) || isGetterOnly) {
-              // If the field was handled by the constructor OR it's a final/getter-only field,
-              // do not attempt to process it further here (especially invokeSetter).
-              return; 
-          }
-
-          // If we reach here, it's a mutable (non-final) flattened field 
-          // that was NOT handled by the constructor. This scenario might be rare
-          // for 'flatten:true' but this logic would apply.
-          final fieldValue = jsonMap.getPropertyValue(property.name); 
-          final objectToDeserialize = meta?.name != null && fieldValue is Map 
-              ? fieldValue.map((key, value) => MapEntry(skipPrefix(meta?.name, key, propertyContext.caseStyle), value)) 
-              : fieldValue;
-          
-          im.invokeSetter(name, _deserializeObject(objectToDeserialize, propertyContext));
-          return; // Ensure we don't fall through to other logic for this property
-      }
-      // END OF BLOCK TO REPLACE/OVERWRITE
-        if (im.invokeGetter(name) == null &&
-            defaultValue != null &&
-            !isGetterOnly) {
-          im.invokeSetter(name, defaultValue);
-        }
-        if (meta?.inject != true) {
-          return;
-        }
-      }
-      fieldValue = property.raw
-          ? _deserializeObject(fieldValue, propertyContext)
-          : property.value;
-      if (isGetterOnly) {
-        if (inheritedPublicFieldNames.contains(name) &&
-            !mappedFields.contains(property.name)) {
-          mappedFields.add(property.name);
-        }
-      } else {
-        // Logic for !isGetterOnly fields:
-        if (meta?.rawJson == true && typeInfo.type == String) {
-          if (fieldValue is Map || fieldValue is List) {
-            // If it's a Map or List, convert to JSON string.
-            fieldValue = JsonEncoder().convert(fieldValue);
-          } else if (fieldValue == null) {
-            // If JSON value is null, field should be null.
-            fieldValue = null;
-          } else {
-            // If it's any other type (e.g., String, num, bool),
-            // convert it to string to ensure type safety for the String field.
-            fieldValue = fieldValue.toString();
-          }
-        }
-
-        // Now, apply decorator and set the property.
-        // fieldValue at this point MUST be a String if rawJson conditions were met and typeInfo.type is String, or null.
-        fieldValue = _applyValueDecorator(fieldValue, typeInfo) ?? defaultValue;
-        im.invokeSetter(name, fieldValue);
-        mappedFields.add(property.name);
-      }
-    });
-
-    final discriminatorPropertyName =
-        _getDiscriminatorProperty(classInfo, context.options);
-    final unmappedFields = jsonMap.map.keys
-        .where((field) =>
-            !mappedFields.contains(field) && field != discriminatorPropertyName)
-        .toList();
-    if (unmappedFields.isNotEmpty) {
-      final jsonAnySetter = classInfo.getJsonAnySetter(context.options.scheme);
-      for (var field in unmappedFields) {
-        final jsonSetter =
-            classInfo.getJsonSetter(field, context.options.scheme) ??
-                jsonAnySetter;
-        final params = jsonSetter == jsonAnySetter
-            ? [field, jsonMap.map[field]]
-            : [jsonMap.map[field]];
-        if (jsonSetter != null) {
-          im.invoke(jsonSetter.simpleName, params);
-        }
-      }
-    }
-
-    return _applyValueDecorator(objectInstance, typeInfo);
   }
 }
